Added Board
- Added Board, which can actually move and correctly respond to user input
- Board can have speed in seconds and move with given speed regardless to framerate

+++ src/models/game_entities/Board/board.c
@@ -14,12 +14,12 @@ void Board_delete(Board* target) {
 	free(target);
 }
 
-void Board_move(Board* target, int delta_x, int delta_y) {
+void Board_move(Board* target, float delta_x, float delta_y) {
 	target->_boardSprite->_spriteFrame.x += delta_x;
 	target->_boardSprite->_spriteFrame.y += delta_y;
 }
 
-void Board_move_to(Board* target, int x, int y) {
+void Board_move_to(Board* target, float x, float y) {
 	target->_boardSprite->_spriteFrame.x = x;
 	target->_boardSprite->_spriteFrame.y = y;
 }
